Update README.md
This is a real-time chatting application built with React for the frontend and Node.js/Express for the backend. It allows users to send messages instantly, with real-time communication powered by Socket.io. The app features user authentication via JWT, secure password hashing with bcryptjs, and media file uploads via Cloudinary. The application is designed to be scalable, secure, and responsive, offering a seamless chatting experience.